+++ progweb-20232/app/Controllers/Http/UserController.ts
@@ -1,68 +1,91 @@
 /* eslint-disable prettier/prettier */
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { CreateUserService } from "../../User/utils";
-import { UserCreateRequestSchema, UserUpdateRequestSchema } from '../../User/interface';
+import { UserCreateRequestSchema, UserType, UserUpdateRequestSchema } from '../../User/interface';
 
 export default class UserController {
     public async create({ request, response }: HttpContextContract) {
         try {
             const newUser = await request.validate({ schema: UserCreateRequestSchema })
             const svc = CreateUserService()
-        
-            const result = svc.createUser(newUser)
-            response.status(201)
-            return result
+            
+            const user = await svc.createUser(newUser as UserType)
+            if (user != undefined) {
+                response.redirect('/auth')
+            }
+            
         } catch (error) {
             return error
         }
     }
 
-    public async update({ params, request, response }: HttpContextContract) {
+    public async update({ params, request, response, auth }: HttpContextContract) {
         try {
-            const svc = CreateUserService()
+            if(auth.isAuthenticated) {
+                const svc = CreateUserService()
 
-            var userToUpdate = await request.validate({ schema: UserUpdateRequestSchema })
-            userToUpdate.id = parseInt(params.id)
-        
-            const result = svc.updateUser(userToUpdate)
-            response.status(200)
-            return result
+                var userToUpdate = await request.validate({ schema: UserUpdateRequestSchema })
+                userToUpdate.id = parseInt(params.id)
+            
+                const result =await svc.updateUser(userToUpdate as UserType)
+                response.status(200)
+                return result
+            } else {
+                response.status(401)
+                response.send("Unauthorized")
+            }
         } catch(error) {
             return error
         }  
     }
 
 
-    public async getByID({ params }: HttpContextContract) {
+    public async getByID({ params, auth, response }: HttpContextContract) {
         try {
-            const svc = CreateUserService()
-            const id = parseInt(params.id);
-    
-            const user = await svc.getUserByID(id)
-            return user
+            if(auth.isAuthenticated) {
+                const svc = CreateUserService()
+                const id = parseInt(params.id);
+        
+                const user = await svc.getUserByID(id)
+                return user
+            } else {
+                response.status(401)
+                response.send("Unauthorized")
+            }
         } catch (error) {
             return error
         }
     }
 
-    public async list({}: HttpContextContract) {
+    public async list({ auth, response }: HttpContextContract) {
         try {
-            const svc = CreateUserService()
+            if(auth.isAuthenticated) {
+                const svc = CreateUserService()
 
-            const users = await svc.listUsers()
-            return users
+                const users = await svc.listUsers()
+                return users
+            } else {
+                response.status(401)
+                response.send("Unauthorized")
+            }
         } catch (error) {
             return error
         }  
     }
 
-    public async deleteByID({ params, response }: HttpContextContract) {
-        try {
-            const svc = CreateUserService()
-            const id = parseInt(params.id)
     
-            await svc.deleteUser(id)
-            response.status(204)
+    public async deleteByID({ params, response, auth }: HttpContextContract) {
+        try {
+            if(auth.isAuthenticated) {
+                const svc = CreateUserService()
+                const id = parseInt(params.id)
+        
+                await svc.deleteUser(id)
+                response.status(204)
+            } else {
+                response.status(401)
+                response.send("Unauthorized")
+            }
         } catch(error) {
             return error
         }
